requirements.txt
# Sign Language Learning App - Backend Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database (Optional - using Supabase)
# flask-sqlalchemy==3.1.1

# Machine Learning
mediapipe==0.10.32
opencv-python==4.9.0.80
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Supabase (uncomment if needed)
# supabase==2.3.0
